resistance bug, sparameter example, devices, sparameter viewer
fixed a bug created in the last checkin where I corrected the spelling
of
resistance in the part property.  This caused all previously generated
schematics (with resistors) to fail.  Just edited all of these previous
schematic files.
Added to the s-parameter example
Added all remaining device pictures to date
Fixed s-parameter viewer so when it is used to view s-parameters of a
device
from the device properties dialog, you can still do all the other
stuff like editing the calculation properties and resampling.