Fixed lots of things and added FourPort tests
added exception for frequency dependent s-parameters in parser
with no frequency provided (old error was too cryptic).
plus test and coverage.
added block solve type to s-parameters numeric - although found
it's really not needed by debugging some other stuff - but
good to have.
updated listing
added condition number check to direct s-parameter computation
method.  This caused some of the diabolical tests to fail
because they threw an exception - which they should have already.
so I added safetees to them and fixed them.
added some use of setUp() to the unit tests to change directories
and to set the adaption strategy - this fixed a bug where the
DesignCon test was failing because it depended on SinX interpolation.