licenses
Added TestHeaders to unit tests that tests all headers to conform
to a good format and contain the proper license.
This script can be used to change the headers (and it was) and
all unit tests and diffs were checked.  From now on the unit tests
test license conformance.
Each file should start with some explanation about the module.  Many
don't and the tests log these, but don't fail the tests.