Don't show cursor values on plots in preferences
a new preference is added that defaults to False
that controls whether the cursor values are shown
on the plots - this is what causes the erratic jumping
complained about by @jsmith04.

while not perfect, it resolves #25
at least until I can figure out
how to keep the plot sizes
fixed.