arrow keys in schematic
added key bindings for using arrow and escape keys in schematic.
currently, you can now move and unselect a part or a group of parts
(which includes a wire) with the keyboard.
Also, when panning is selected, you can move the drawings origin
with the arrow keys, and stop panning with the escape key.