resolved directory issues (I think) with external projects
They were not being opened properly if they were not in same directory.
Now using absolute path.
added --external argument so that the externally opened projects do not
remember which files were opened.  Previously, an annoying thing was
that if
you opened an external project, then when you closed the main project
and reopened it, the external project would open.
This was solved by not keeping track of opened files at all when
projects
are opened externally.