Tkinter slowness improvements
previously, I had made lots of changes to, when a window is closed,
withdraw the dialog so that the user does not see the window being
destroyed piece by piece.  This is especially true for windows
containing matplotlib plots, which delete slowly and painfully.
This call was not being made when a window was destroyed by the
root window being destroyed, so I overload the tk destroy function
to withdraw the window invariably prior to destruction.