Fixed bug - spaces in filenames
filenames with spaces can be quoted (single-quotes).
The SignalIntegrityApp now quotes filenames in the netlist if a space
is found in them.
This is experimental for now, because I made a risky change in
DeviceParser, which split and rejoined argument lists on space
boundaries.
I don't know the implications yet of removing that behavior.
Passes all unit tests after some fixes.
I'm not sure whether subcircuit filenames can have
spaces in them after this fix - but it's still an
improvement.
fixes #28