Added known s-parameters dictonary feature
Added a feature whereby the parser classes (which all derive from
SystemDescriptionParser) have the capability to provide a dictionary
of known devices.  This dictionary has a key value equal to the
device netlist line (following the device keyword and reference
designator) and the value is the s-parameters.

This allows for using the parser classes with devices declared that
are weird or don't have to be read from the file system.  In the
past, usually one would write s-parameters to a file and then
use that in a netlist as a file device.  Now, one could write
and netlist line like 'device D1 2 radiotube' and supply the dictionary
{'2 radiotube':sparameters} where s-parameters are the two-port
s-parameters of a radiotube.

todo:  add this usage to the system device declaration in deembedding