step and pulse waveform fixes (adopt some legacy behavior)
The step (and pulse) risetime defaulting to 1/Ts caused lots of tests to
fail that depended on legacy behavior.  Therefore, it defaults to
zero and actually is zero.  If you specify something other than zero
then it crosses the 50% point at the start time.
I'll have to revisit this behavior at some point to see if this is what
I actually want.
Also, updated the code documentation.
closes #29