Net Names
regards issue #38

Added a net name device that allows it to act like a net name
in that any wire or part connected to it with the same net name
property value has their networks connected together.

In the process, I did remove the reference property from the port
and this net name device, which is a bit risky.  I did run all tests.

The only improvements that might be made is to remove the dot where
the net name is joined to the wire (I don't love it but it removes
ambiguity) and to test whether two different named net name devices
ever get connected together. (That would be the behavior of most
tools, but I feel that it is harmless).

This feature does not affect the text netlist that gets created.

All Python2 and Python3 tests were run and pass.