fixed showing impedance in transfer parameters
As transfer parameters share the s-parameter viewer, it was seen that
when disabled views like show impedance, excess inductance, and
excess capacitance were selected in the preferences and the transfer
parameters were viewed, although the settings were grayed out, they
were being used anyway.

The fix was to check both the setting and whether the setting is
activated or not to determine how plots are made.