fixed some wrong activations
I was actually setting some preferences based on the type of parameter
being viewed.  This used to be okay when they were not held in
the preferences, but now, this causes the preferences to change
when viewing different things, which is wrong.
I fixed this for causality, passivity, impedance, etc.
but still need to fix it for some of the zoom joins, which I
don't know what to do about yet.