ask open file - prefer extension of file already being used
when files are opened, there are multiple extension choices.  For
example, s-parameter files could be projects.  In the past, we
always prefer the s-parameter file, but now, if there is already
a project file in the entry, it will browse first for other project
files.  This makes things less frustrating.