Features preference
Added feature category in preferences and the NetworkAnalyzerModel
feature.

The purpose is to sort of hide the network analyzer model devices
so that the general user does not need to see these, and so I can
avoid documentating this feature, which is useful to only a
select few users.  The feature is enabled by manually editing
the preferences .xml file.