Integrated LeCroy Waveform into Application
Now, whenever you read or write a waveform file, if it is .trc,
then it is read or written as LeCroy format.
Also, waveforms for sources can use .trc, and when you write
simulation waveforms, they can be written as .trc.
A preference was added to determine whether the user prefers
LeCroy format files so that when waveforms are saved, the proper
preference is given (although one could still choose the other
format).