Performance improvement
I discovered that the impedance peeling on ports is being done all
the time, even if the times specified are zero.  This really slowed
things down.  Now, if they are all zero, it does none of it, and
even when not all ports are peeled, the unpeeled ports don't do
the expensive impedance profile calculation.