Dramaticaly improved Eye Diagrams
Introduced enhanced precision.  This introduces a step size,
that allows for a single line from a pixel to another to be
broken down in smaller steps, after which the probability of
hitting adjacent pixels is calculated.  As such, a pixel size
spot is swept across the bit map.
This especially helps when the waveform is slewing multiple
rows across the eye diagram - this fills in those rows with the
correct probabilities.
In an Auto mode, the number of steps depends on the difference
between the rows of the pixels -- this allows for balancing
between the calculation requirements and the quality of the eye
in high slew rate regions.