Abort if file not saved
When closing or when opening a new file or creating
a new project, the user is asked if he/she wants to
save the project.  In the past, if he said yes, but
failed to pick a file, the project was closed anyway.

The new behavior is to abort the action if the file
was not saved.